fixes #2 regarding compile (#8)
* fixes #2 regarding compile

when using shortened code in test2.cpp:

```C++
optparse::OptionParserExcept parser = optparse::OptionParserExcept().version(version)
           .description("Description")
           .epilog("Epilog");
```

The compiler errors with:

> error: conversion from 'optparse::OptionParser' to non-scalar type 'optparse::OptionParserExcept' requested

That's because `optparse::OptionParserExcept` doesn't define a copy constructor
of the Base class. Defining this, removes the error and allows it function as
intended.

* fixes #2 regarding compile

when using shortened code in test2.cpp:

```cpp
optparse::OptionParserExcept parser = optparse::OptionParserExcept().version(version)
           .description("Description")
           .epilog("Epilog");
```

The compiler errors with:

> error: conversion from 'optparse::OptionParser' to non-scalar type 'optparse::OptionParserExcept' requested

That's because `optparse::OptionParserExcept` doesn't define a copy constructor
of the Base class. Defining this, removes the error and allows it function as
intended.